Add dockerfile for building zokrates_js, modify ci steps
