+++ .circleci/config.yml
@@ -104,7 +104,7 @@ jobs:
           command: ZOKRATES_HOME=$(pwd)/zokrates_stdlib/stdlib/ WITH_LIBSNARK=1 RUSTFLAGS="-D warnings" ./full_test.sh
   deploy:
     docker:
-      - image: circleci/python
+      - image: circleci/python:latest-node
     steps:
       - checkout
       - setup_remote_docker:
@@ -112,6 +112,21 @@ jobs:
       - run:
           name: Release
           command: ./scripts/release.sh
+  zokrates_js_build:
+    docker:
+      - image: circleci/node
+    steps:
+      - checkout
+      - setup_remote_docker
+      - run:
+          command: ./zokrates_js/build.sh
+  zokrates_js_test:
+    docker:
+      - image: circleci/node
+    steps:
+      - checkout
+      - run:
+          command: cd zokrates_js && npm run test
 workflows:
   version: 2
   build-test-and-deploy:
@@ -120,6 +135,8 @@ workflows:
       - test
       - wasm_test
       - integration_test
+      - zokrates_js_build
+      - zokrates_js_test
       - deploy:
           filters:
             branches:
@@ -130,3 +147,5 @@ workflows:
             - test
             - wasm_test
             - integration_test
+            - zokrates_js_build
+            - zokrates_js_test
