+++ zokrates_js/build.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e
+
+docker build -t zokrates_js -f zokrates_js/Dockerfile .
