+++ zokrates_js/publish.sh
@@ -1,12 +1,5 @@
 #!/bin/bash
 
-CWD="${0%/*}"
-if [ "$0" != "$CWD" ] && [ "$CWD" != "" ]; then 
-    cd $CWD
-fi
-
-NPM_VERSION=$(npm view zokrates-js dist-tags.latest)
-
 PACKAGE_VERSION=$(cat package.json \
   | grep version \
   | head -1 \
@@ -19,19 +12,18 @@ CARGO_VERSION=$(cat Cargo.toml \
   | awk '{print $3}' \
   | sed -e 's/"//g')
 
-if [ $NPM_VERSION = $PACKAGE_VERSION ]; then
-    echo "Latest npm version is equal to current package version. Up the version to publish to npm."
-    exit 0
-fi
-
 if [ $PACKAGE_VERSION != $CARGO_VERSION ]; then
     echo "Cargo crate version must be equal to npm package version ($CARGO_VERSION -> $PACKAGE_VERSION)"
     exit 0
 fi
 
-# npm
-npm run build
+NPM_VERSION=$(npm view zokrates-js dist-tags.latest)
 
-# publish package
+if [ $NPM_VERSION = $PACKAGE_VERSION ]; then
+    echo "Latest npm version is equal to current package version. Up the version to publish to npm."
+    exit 0
+fi
+
+# publish
 npm set //registry.npmjs.org/:_authToken=${NPM_TOKEN}
 npm publish
